added test angle

+++ PolygonChecker/triangleSolver.h
@@ -2,5 +2,5 @@
 
 void getAngles(float a, float b, float c, float arr[]);
 char* analyzeTriangle(int side1, int side2, int side3);
-double radiansToDegrees(double radians);
+float radiansToDegrees(float radians);
